fix: too much rerendering when searching for a specific note
